requirements.txt
python-podcastindex>=1.15.0
feedparser>=6.0.10
pyairtable>=3.0.0
requests>=2.28.0
python-dotenv>=1.0.0
anthropic>=0.50.0
beautifulsoup4>=4.12.0
ollama>=0.4.0
pydantic>=2.0.0
